requirements.txt
xmpppy
pydantic
python-dotenv==0.21.0
fastapi==0.82.0
uvicorn

